+++ src/main/java/oshi/software/os/windows/WindowsOperatingSystem.java
@@ -20,7 +20,6 @@ public class WindowsOperatingSystem implements OperatingSystem {
 	
 	private OperatingSystemVersion _version = null;
 	
-	@Override
 	public OperatingSystemVersion getVersion() {
 		if (_version == null) {
 			_version = new OSVersionInfoEx();
@@ -28,17 +27,14 @@ public class WindowsOperatingSystem implements OperatingSystem {
 		return _version;
 	}
 
-	@Override
 	public String getFamily() {
 		return "Windows";
 	}
 
-	@Override
 	public String getManufacturer() {
 		return "Microsoft";
 	}
 	
-	@Override
 	public String toString() {
 		StringBuilder sb = new StringBuilder();
 		sb.append(getManufacturer());