+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Next
 ====
 
+* [#5](https://github.com/dblock/oshi/pull/5): Added Linux support - [@ptitvert](https://github.com/ptitvert).
 * [#3](https://github.com/dblock/oshi/pull/3): Mavenized project - [@le-yams](https://github.com/le-yams).
 
 1.1 (10/13/2013)